Bug fixed
The issue was that the wasRun flag was set to true by `compileTest` but
not reset to false by `runTest`. This caused the test falling not being
recognized because the flag was already set to true. The fix was to
create different flags for every processus (4 in total). This allows to
have different states for each processus.